+++ src/scanners/__init__.py
@@ -4,5 +4,6 @@ from .s3_scanner import S3Scanner
 from .ec2_scanner import EC2Scanner
 from .vpc_scanner import VPCScanner
 from .rds_scanner import RDSScanner
+from .lambda_scanner import LambdaScanner
 
-__all__ = ["BaseScanner", "IAMScanner", "S3Scanner", "EC2Scanner", "VPCScanner", "RDSScanner"]
+__all__ = ["BaseScanner", "IAMScanner", "S3Scanner", "EC2Scanner", "VPCScanner", "RDSScanner", "LambdaScanner"]